requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.3.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
dulwich==0.21.7
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==24.3.1
magic-filter==1.0.12
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
openapi-python-client==0.22.0
packaging==24.2
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
propcache==0.2.1
psycopg2==2.9.10
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.8.4
SecretStorage==3.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.0
yarl==1.18.3
